perf_hooks: fix truncation of monitorEventLoopDelay() resolution
`IntervalHistogram` stored the interval as `int32_t`, so a resolution
above 2^31 - 1 ms wrapped: `resolution: 2 ** 32 + 1` sampled every
millisecond.

+++ src/histogram.cc
@@ -2616,7 +2616,7 @@ void IntervalHistogram::RegisterExternalReferences(
 IntervalHistogram::IntervalHistogram(Environment* env,
                                      Local<Object> wrap,
                                      AsyncWrap::ProviderType type,
-                                     int32_t interval,
+                                     uint64_t interval,
                                      OnInterval on_interval,
                                      const Histogram::Options& options)
     : HandleWrap(env, wrap, reinterpret_cast<uv_handle_t*>(&timer_), type),
@@ -2633,7 +2633,7 @@ IntervalHistogram::IntervalHistogram(Environment* env,
 
 BaseObjectPtr<IntervalHistogram> IntervalHistogram::Create(
     Environment* env,
-    int32_t interval,
+    uint64_t interval,
     OnInterval on_interval,
     const Histogram::Options& options,
     AsyncWrap::ProviderType type) {
